requirements.txt
# Streamlit version requirements
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0

# Existing requirements
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
feedparser>=6.0.10
apscheduler>=3.10.4
jinja2>=3.1.2
python-multipart>=0.0.6